Перенес фетч и изменение стора в отдельный namespace
