+++ project/src/components/pages/favorites-page/favorites-page.tsx
@@ -1,11 +1,11 @@
 import { useEffect } from 'react';
 
 import { useAppDispatch, useAppSelector } from 'src/hooks';
-import { fetchFavorites } from 'src/store/user-process/api-actions';
 import { switchFavoritesPageComponent } from 'src/components/pages/favorites-page/helpers/switch-favorites-page-component';
+import { fetchFavorites } from 'src/store/favorite-page-process/api-actions';
 
 function FavoritesPage() {
-  const favorites = useAppSelector((state) => state.USER.favorites);
+  const favorites = useAppSelector((state) => state.FAVORITE_PAGE.favorites);
   const dispatch = useAppDispatch();
 
   useEffect(() => {
