+++ project/src/components/pages/favorites-page/helpers/get-favorites-state.ts
@@ -1,6 +1,6 @@
-import { FavoritesState } from 'src/types/user-process';
 import { GroupedCity } from 'src/types/main-page';
 import { isEmpty } from 'src/components/pages/favorites-page/helpers/is-empty';
+import { FavoritesState } from 'src/types/favorite-process';
 
 export const getFavoritesState = (
   favorites: FavoritesState.Unknown | GroupedCity[],
