+++ project/src/components/pages/favorites-page/helpers/is-empty.ts
@@ -1,5 +1,5 @@
-import { FavoritesState } from 'src/types/user-process';
 import { GroupedCity } from 'src/types/main-page';
+import { FavoritesState } from 'src/types/favorite-process';
 
 export const isEmpty = (favorites: FavoritesState.Unknown | GroupedCity[]): boolean =>
   favorites !== 'unknown' && !favorites.length;
