+++ project/src/components/pages/favorites-page/helpers/switch-favorites-page-component.tsx
@@ -1,9 +1,9 @@
-import { FavoritesState } from 'src/types/user-process';
 import { GroupedCity } from 'src/types/main-page';
 import Loader from 'src/components/shared/loader/loader';
 import FavoritesPageEmpty from 'src/components/pages/favorites-page/favorites-page-empty';
 import FavoritesPageContent from 'src/components/pages/favorites-page/favorites-page-content';
 import { getFavoritesState } from 'src/components/pages/favorites-page/helpers/get-favorites-state';
+import { FavoritesState } from 'src/types/favorite-process';
 
 export const switchFavoritesPageComponent = (
   favorites: FavoritesState.Unknown | GroupedCity[],
