+++ project/src/components/shared/bookmark/bookmark.tsx
@@ -4,7 +4,11 @@ import { BookmarkProps } from 'src/types/bookmark';
 import { getAriaText } from 'src/components/shared/bookmark/helpers/get-aria-text';
 import { toggleActiveBookmarkClass } from 'src/components/shared/bookmark/helpers/toogle-active-bookmark-class';
 
-function Bookmark({ isActive: inBookmark, onToggleActive = () => undefined }: BookmarkProps) {
+function Bookmark({
+  isActive: inBookmark,
+  isLoading = false,
+  onToggleActive = () => undefined,
+}: BookmarkProps) {
   const [isActive, setIsActive] = useState(inBookmark);
 
   const handleBookmarkClick = () => {
@@ -20,6 +24,7 @@ function Bookmark({ isActive: inBookmark, onToggleActive = () => undefined }: Bo
       onClick={handleBookmarkClick}
       className={toggleActiveBookmarkClass({ isActive })}
       type="button"
+      disabled={isLoading}
     >
       <svg className="place-card__bookmark-icon" width={18} height={19}>
         <use xlinkHref="#icon-bookmark" />
