+++ project/src/store/constants/constants.ts
@@ -1,4 +1,5 @@
 export enum NameSpace {
   MainPage = 'MAIN_PAGE',
   User = 'USER',
+  FavoritePage = 'FAVORITE_PAGE',
 }
