+++ project/src/store/favorite-page-process/action-type.ts
@@ -1,0 +1,4 @@
+export enum ActionType {
+  SetFavorites = 'setFavorites',
+  FetchFavorites = 'fetchFavorites',
+}
