+++ project/src/store/user-process/action-type.ts
@@ -1,7 +1,5 @@
 export enum ActionType {
   SetAuthStatus = 'setAuthStatus',
-  SetFavorites = 'setFavorites',
-  FetchFavorites = 'fetchFavorites',
   SetMe = 'setMe',
   Login = 'login',
   Logout = 'logout',
