+++ project/src/store/user-process/user-process.ts
@@ -1,16 +1,13 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-import { FavoritesState, InitialState } from 'src/types/user-process';
 import { AuthorizationStatus } from 'src/types/auth';
 import { NameSpace } from 'src/store/constants/constants';
 import { ActionType } from 'src/store/user-process/action-type';
 import { User } from 'src/types/user';
-import { groupCities } from 'src/helpers/group-cities';
-import { Favorite } from 'src/types/favorite';
+import { InitialState } from 'src/types/user-process';
 
 const initialState: InitialState = {
   authorizationStatus: AuthorizationStatus.Unknown,
-  favorites: FavoritesState.Unknown,
   me: undefined,
 };
 
@@ -27,10 +24,7 @@ export const userProcess = createSlice({
     [ActionType.SetMe](state, { payload: user }: { payload: User | undefined }) {
       state.me = user;
     },
-    [ActionType.SetFavorites](state, { payload: favorites }: { payload: Favorite[] }) {
-      state.favorites = groupCities(favorites);
-    },
   },
 });
 
-export const { setAuthStatus, setMe, setFavorites } = userProcess.actions;
+export const { setAuthStatus, setMe } = userProcess.actions;
