+++ project/src/types/bookmark.ts
@@ -1,4 +1,5 @@
 export type BookmarkProps = {
   isActive: boolean;
+  isLoading?: boolean;
   onToggleActive?: (isActive: boolean) => void;
 };
